+++ lib/mbed-connector.js
@@ -234,7 +234,6 @@ mbedClient.prototype.startLongPolling = function(callback, options) {
   function poll(error) {
     if (error) {
       console.log('ERROR: Long poll failed [Status ' + error.status + ']');
-      console.log(error);
       throw error;
     } else {
       _this.longPollRequestObj = _this.makeAuthorizedRequest(options, poll);
